refactor: create tree test
