+++ crates/committer/src/patricia_merkle_tree/original_skeleton_tree/create_tree_test.rs
@@ -1,13 +1,11 @@
 use super::OriginalSkeletonTreeImpl;
-use crate::block_committer::input::StarknetStorageValue;
 use crate::felt::Felt;
 use crate::hash::hash_trait::HashOutput;
-use crate::patricia_merkle_tree::filled_tree::node::{ClassHash, CompiledClassHash, Nonce};
 use crate::patricia_merkle_tree::internal_test_utils::OriginalSkeletonMockTrieConfig;
 use crate::patricia_merkle_tree::internal_test_utils::{small_tree_index_to_full, MockLeaf};
 use crate::patricia_merkle_tree::node_data::inner_node::EdgePath;
 use crate::patricia_merkle_tree::node_data::inner_node::{EdgePathLength, PathToBottom};
-use crate::patricia_merkle_tree::node_data::leaf::{ContractState, LeafModifications};
+use crate::patricia_merkle_tree::node_data::leaf::LeafModifications;
 use crate::patricia_merkle_tree::original_skeleton_tree::create_tree::SubTree;
 use crate::patricia_merkle_tree::original_skeleton_tree::node::OriginalSkeletonNode;
 use crate::patricia_merkle_tree::original_skeleton_tree::tree::OriginalSkeletonTree;
@@ -420,26 +418,6 @@ pub(crate) fn create_32_bytes_entry(simple_val: u128) -> [u8; 32] {
 pub(crate) fn create_mock_leaf_entry(val: u128) -> (StorageKey, StorageValue) {
     let leaf = MockLeaf(Felt::from(val));
     (leaf.get_db_key(&leaf.0.to_bytes_be()), leaf.serialize())
-}
-
-pub(crate) fn create_storage_leaf_entry(val: u128) -> (StorageKey, StorageValue) {
-    let leaf = StarknetStorageValue(Felt::from(val));
-    (leaf.get_db_key(&leaf.0.to_bytes_be()), leaf.serialize())
-}
-
-pub(crate) fn create_compiled_class_leaf_entry(val: u128) -> (StorageKey, StorageValue) {
-    let leaf = CompiledClassHash(Felt::from(val));
-    (leaf.get_db_key(&leaf.0.to_bytes_be()), leaf.serialize())
-}
-
-pub(crate) fn create_contract_state_leaf_entry(val: u128) -> (StorageKey, StorageValue) {
-    let felt = Felt::from(val);
-    let leaf = ContractState {
-        nonce: Nonce(felt),
-        storage_root_hash: HashOutput(felt),
-        class_hash: ClassHash(felt),
-    };
-    (leaf.get_db_key(&felt.to_bytes_be()), leaf.serialize())
 }
 
 fn create_patricia_key(val: u128) -> StorageKey {
